src/portfolioagent/requirements.txt
grpcio
grpcio-tools
protobuf
PyJWT
cryptography